+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/inspections/GitlabLintInspection.kt
@@ -1,15 +1,13 @@
 package com.github.blarc.gitlab.template.lint.plugin.inspections
 
-import com.github.blarc.gitlab.template.lint.plugin.GitlabLintBundle.message
-import com.github.blarc.gitlab.template.lint.plugin.GitlabLintUtils.matchesGitlabLintRegex
 import com.github.blarc.gitlab.template.lint.plugin.runLinting
+import com.github.blarc.gitlab.template.lint.plugin.settings.AppSettings
+import com.github.blarc.gitlab.template.lint.plugin.ui.settings.LintFrequencyEnum
 import com.intellij.codeInspection.LocalInspectionTool
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.openapi.components.Service
-import com.intellij.openapi.progress.runBackgroundableTask
 import com.intellij.psi.PsiElementVisitor
 import com.intellij.psi.PsiFile
-import com.intellij.ui.EditorNotifications
 
 @Service
 class GitlabLintInspection : LocalInspectionTool() {
@@ -17,13 +15,8 @@ class GitlabLintInspection : LocalInspectionTool() {
     override fun buildVisitor(holder: ProblemsHolder, isOnTheFly: Boolean): PsiElementVisitor {
         return object : PsiElementVisitor() {
             override fun visitFile(file: PsiFile) {
-                runBackgroundableTask(message("inspection.title"), file.project) { indicator ->
-                    indicator.isIndeterminate = true
-                    val project = file.project
-                    if (matchesGitlabLintRegex(file.name)) {
-                        runLinting(project, file)
-                    }
-                    EditorNotifications.getInstance(project).updateAllNotifications()
+                if (AppSettings.instance?.lintFrequency == LintFrequencyEnum.ON_CHANGE) {
+                    runLinting(file)
                 }
             }
         }